+++ trace-tools/trace-tools/CHANGELOG.md
@@ -24,6 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  - Initial crate:
-   - Flamegraph and logging layers
-   - Subscriber builder
-   - `Observe` trait
+   - Flamegraph and logging layers;
+   - Subscriber builder;
+   - `Observe` trait;